refactor: To make it easier to test individual components
- Move findAllBeacons to a separate package to rule out tests interfering with each other.

- Remove HATEOAS as that seems to have causing issues. Will add in a future commit.
