+++ service/src/main/java/uk/gov/mca/beacons/api/beaconsearch/rest/BeaconSearchController.java
@@ -1,20 +1,17 @@
-package uk.gov.mca.beacons.api.search.beacons.rest;
-
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
+package uk.gov.mca.beacons.api.beaconsearch.rest;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PagedResourcesAssembler;
-import org.springframework.hateoas.EntityModel;
-import org.springframework.hateoas.PagedModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import uk.gov.mca.beacons.api.search.BeaconSearchService;
+import uk.gov.mca.beacons.api.beaconsearch.BeaconSpecificationSearchService;
 import uk.gov.mca.beacons.api.search.domain.BeaconSearchEntity;
 
 @Slf4j
@@ -23,23 +20,22 @@ import uk.gov.mca.beacons.api.search.domain.BeaconSearchEntity;
 @Tag(name = "Find All Beacons")
 public class BeaconSearchController {
 
-  private final BeaconSearchService beaconSearchService;
+  private final BeaconSpecificationSearchService beaconSpecificationSearchService;
   private final PagedResourcesAssembler<BeaconSearchEntity> pagedAssembler;
 
+  @Autowired
   public BeaconSearchController(
-    BeaconSearchService beaconSearchService,
+    BeaconSpecificationSearchService beaconSpecificationSearchService,
     PagedResourcesAssembler<BeaconSearchEntity> pagedAssembler
   ) {
-    this.beaconSearchService = beaconSearchService;
+    this.beaconSpecificationSearchService = beaconSpecificationSearchService;
     this.pagedAssembler = pagedAssembler;
   }
 
   @Cacheable(value = "find-all-beacons")
   @GetMapping("/find-all")
   @Operation(summary = "Find all beacons matching specific fields (paginated)")
-  public ResponseEntity<
-    PagedModel<EntityModel<BeaconSearchEntity>>
-  > findAllBeacons(
+  public ResponseEntity<Page<BeaconSearchEntity>> findAllBeacons(
     @RequestParam(required = false, defaultValue = "") String status,
     @RequestParam(required = false, defaultValue = "") String uses,
     @RequestParam(required = false, defaultValue = "") String hexId,
@@ -55,45 +51,22 @@ public class BeaconSearchController {
     Pageable pageable
   ) {
     try {
-      Page<BeaconSearchEntity> results = beaconSearchService.findAllBeacons(
-        status,
-        uses,
-        hexId,
-        ownerName,
-        cospasSarsatNumber,
-        manufacturerSerialNumber,
-        pageable
-      );
+      Page<BeaconSearchEntity> results =
+        beaconSpecificationSearchService.findAllBeacons(
+          status,
+          uses,
+          hexId,
+          ownerName,
+          cospasSarsatNumber,
+          manufacturerSerialNumber,
+          pageable
+        );
 
-      var pagedModel = pagedAssembler.toModel(
-        results,
-        beacon ->
-          EntityModel.of(
-            beacon,
-            linkTo(BeaconSearchController.class)
-              .slash(beacon.getId())
-              .withSelfRel(),
-            linkTo(BeaconSearchController.class)
-              .slash(beacon.getId())
-              .withRel("beaconSearchEntity")
-          ),
-        linkTo(
-          methodOn(BeaconSearchController.class).findAllBeacons(
-            status,
-            uses,
-            hexId,
-            ownerName,
-            cospasSarsatNumber,
-            manufacturerSerialNumber,
-            pageable
-          )
-        ).withRel("beaconSearch")
-      );
-      return ResponseEntity.ok(pagedModel);
+      return ResponseEntity.ok(results);
     } catch (Exception ex) {
       log.error("Failed to fetch beacons", ex);
       return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(
-        PagedModel.empty()
+        Page.empty()
       );
     }
   }
