+++ service/src/test/java/uk/gov/mca/beacons/api/search/rest/BeaconSearchRestRepositoryIntegrationTest.java
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.web.reactive.function.BodyInserters;
-import uk.gov.mca.beacons.api.RetryUtils;
 import uk.gov.mca.beacons.api.WebIntegrationTest;
 
 class BeaconSearchRestRepositoryIntegrationTest extends WebIntegrationTest {
@@ -222,194 +221,6 @@ class BeaconSearchRestRepositoryIntegrationTest extends WebIntegrationTest {
     }
   }
 
-  @Nested
-  class GetFindAllBeaconsResults {
-
-    private static final String FIND_ALL_BEACONS_URI =
-      "/spring-api/search/beacons/find-all";
-
-    @Test
-    void shouldFindTheLegacyBeaconByHexIdStatusAndUses() throws Throwable {
-      final var randomHexId = UUID.randomUUID().toString();
-      createLegacyBeacon(randomHexId);
-
-      RetryUtils.runWithRetries(() ->
-        webTestClient
-          .get()
-          .uri(uriBuilder ->
-            uriBuilder
-              .path(FIND_ALL_BEACONS_URI)
-              .queryParam("status", "migrated")
-              .queryParam("uses", "maritime")
-              .queryParam("hexId", randomHexId)
-              .queryParam("ownerName", "")
-              .queryParam("cospasSarsatNumber", "")
-              .queryParam("manufacturerSerialNumber", "")
-              .build()
-          )
-          .exchange()
-          .expectStatus()
-          .isOk()
-          .expectBody()
-          .jsonPath("_embedded.beaconSearch[0].hexId")
-          .isEqualTo(randomHexId)
-          .jsonPath("page.totalElements")
-          .isEqualTo(1)
-      );
-    }
-
-    @Test
-    void shouldFindTheCreatedBeaconByHexIdStatusAndUses() throws Throwable {
-      final String accountHolderId = seedAccountHolder();
-      final var randomHexId = UUID.randomUUID().toString();
-      createBeacon(randomHexId, accountHolderId);
-
-      RetryUtils.runWithRetries(() ->
-        webTestClient
-          .get()
-          .uri(uriBuilder ->
-            uriBuilder
-              .path(FIND_ALL_BEACONS_URI)
-              .queryParam("status", "new")
-              .queryParam("uses", "fishing vessel")
-              .queryParam("hexId", randomHexId)
-              .queryParam("ownerName", "")
-              .queryParam("cospasSarsatNumber", "")
-              .queryParam("manufacturerSerialNumber", "")
-              .build()
-          )
-          .exchange()
-          .expectStatus()
-          .isOk()
-          .expectBody()
-          .jsonPath("_embedded.beaconSearch[0].hexId")
-          .isEqualTo(randomHexId)
-          .jsonPath("page.totalElements")
-          .isEqualTo(1)
-      );
-    }
-
-    @Test
-    void shouldFindTheCreatedLegacyBeaconWithAllFiltersSet() throws Throwable {
-      var legacyBeaconFixtureHexId = "9D0E1D1B8C00001";
-      var legacyBeaconFixtureOwnerName = "Mr Beacon";
-      var legacyBeaconFixtureCospasSarsatNumberValue = 476899;
-      var legacyBeaconFixtureManufacturerSerialNumber =
-        "manufacturer_serial_number_value";
-
-      var random = new Random();
-      var uniqueLegacyBeaconHexId = UUID.randomUUID().toString();
-      var uniqueLegacyBeaconOwnerName = UUID.randomUUID().toString();
-      var pseudoUniqueLegacyBeaconCospasSarsatNumber = random.nextInt(
-        Integer.MAX_VALUE
-      );
-      var uniqueLegacyBeaconManufacturerSerialNumber = UUID.randomUUID()
-        .toString();
-
-      createLegacyBeacon(request ->
-        request
-          .replace(legacyBeaconFixtureHexId, uniqueLegacyBeaconHexId)
-          .replace(legacyBeaconFixtureOwnerName, uniqueLegacyBeaconOwnerName)
-          .replace(
-            Integer.toString(legacyBeaconFixtureCospasSarsatNumberValue),
-            Integer.toString(pseudoUniqueLegacyBeaconCospasSarsatNumber)
-          )
-          .replace(
-            legacyBeaconFixtureManufacturerSerialNumber,
-            uniqueLegacyBeaconManufacturerSerialNumber
-          )
-      );
-
-      RetryUtils.runWithRetries(() ->
-        webTestClient
-          .get()
-          .uri(uriBuilder ->
-            uriBuilder
-              .path(FIND_ALL_BEACONS_URI)
-              .queryParam("status", "MIGRATED")
-              .queryParam("uses", "MARITIME")
-              .queryParam("hexId", uniqueLegacyBeaconHexId)
-              .queryParam("ownerName", uniqueLegacyBeaconOwnerName)
-              .queryParam(
-                "cospasSarsatNumber",
-                pseudoUniqueLegacyBeaconCospasSarsatNumber
-              )
-              .queryParam(
-                "manufacturerSerialNumber",
-                uniqueLegacyBeaconManufacturerSerialNumber
-              )
-              .build()
-          )
-          .exchange()
-          .expectStatus()
-          .isOk()
-          .expectBody()
-          .jsonPath("page.totalElements")
-          .isEqualTo(1)
-          .jsonPath("_embedded.beaconSearch[0].hexId")
-          .isEqualTo(uniqueLegacyBeaconHexId)
-      );
-    }
-
-    @Test
-    void shouldFindTheHATEOASLink() throws Throwable {
-      final String accountHolderId = seedAccountHolder();
-      final var randomHexId = UUID.randomUUID().toString();
-      createBeacon(randomHexId, accountHolderId);
-
-      RetryUtils.runWithRetries(() ->
-        webTestClient
-          .get()
-          .uri(uriBuilder ->
-            uriBuilder
-              .path(FIND_ALL_BEACONS_URI)
-              .queryParam("status", "new")
-              .queryParam("uses", "fishing vessel")
-              .queryParam("hexId", randomHexId)
-              .queryParam("ownerName", "")
-              .queryParam("cospasSarsatNumber", "")
-              .queryParam("manufacturerSerialNumber", "")
-              .build()
-          )
-          .exchange()
-          .expectStatus()
-          .isOk()
-          .expectBody()
-          .jsonPath("_links.self.href")
-          .isNotEmpty()
-      );
-    }
-
-    @Test
-    void shouldFindTheHATEOASEntityLink() throws Throwable {
-      final String accountHolderId = seedAccountHolder();
-      final var randomHexId = UUID.randomUUID().toString();
-      createBeacon(randomHexId, accountHolderId);
-
-      RetryUtils.runWithRetries(() ->
-        webTestClient
-          .get()
-          .uri(uriBuilder ->
-            uriBuilder
-              .path(FIND_ALL_BEACONS_URI)
-              .queryParam("status", "new")
-              .queryParam("uses", "fishing vessel")
-              .queryParam("hexId", randomHexId)
-              .queryParam("ownerName", "")
-              .queryParam("cospasSarsatNumber", "")
-              .queryParam("manufacturerSerialNumber", "")
-              .build()
-          )
-          .exchange()
-          .expectStatus()
-          .isOk()
-          .expectBody()
-          .jsonPath("_embedded.beaconSearch[0]._links.beaconSearchEntity.href")
-          .isNotEmpty()
-      );
-    }
-  }
-
   private String readFile(String filePath) throws Exception {
     return Files.readString(Paths.get(filePath));
   }
@@ -432,17 +243,6 @@ class BeaconSearchRestRepositoryIntegrationTest extends WebIntegrationTest {
 
   private void createLegacyBeacon(String hexId) throws Exception {
     createLegacyBeacon(request -> request.replace("9D0E1D1B8C00001", hexId));
-  }
-
-  private void createLegacyBeaconWithManufacturerSerialNumber(
-    String hexId,
-    String manufacturerSerialNumber
-  ) throws Exception {
-    createLegacyBeacon(request ->
-      request
-        .replace("9D0E1D1B8C00001", hexId)
-        .replace("manufacturer_serial_number_value", manufacturerSerialNumber)
-    );
   }
 
   private void createBeacon(String hexId, String accountHolderId)
